requirements.txt
babel==2.17.0
backrefs==5.9
beautifulsoup4==4.13.4
Brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
csscompressor==0.9.5
cssselect2==0.8.0
fonttools==4.59.1
ghp-import==2.1.0
greenlet==3.2.4
htmlmin2==0.1.13
idna==3.10
Jinja2==3.1.6
jsmin==3.0.1
libsass==0.23.0
livereload==2.7.1
lxml==6.0.1
Markdown==3.8.2
MarkupSafe==3.0.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-exporter==6.2.0
mkdocs-extra-sass-plugin==0.1.0
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.18
mkdocs-material-extensions==1.3.1
mkdocs-minify-plugin==0.8.0
mkdocs-with-pdf==0.9.3
packaging==25.0
paginate==0.5.7
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
playwright==1.54.0
pycparser==2.22
pydyf==0.11.0
pyee==13.0.0
Pygments==2.19.2
pymdown-extensions==10.16.1
pypandoc==1.15
pypdf==6.0.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
PyYAML==6.0.2
pyyaml_env_tag==1.1
requests==2.32.5
six==1.17.0
soupsieve==2.7
tinycss2==1.4.0
tinyhtml5==2.0.0
tornado==6.5.2
typing_extensions==4.14.1
urllib3==2.5.0
watchdog==6.0.0
weasyprint==66.0
webencodings==0.5.1
zopfli==0.2.3.post1
